add zfs support
apparently zfs is broken in 6.2 due to some symbols having changed to
GPL only. We can ignore errors caused by this by saying the ZFS modules
are GPL-licensed in their META file. It should be legal to build against
the GPL-only symbols as long as you don't redistribute the resulting
module binaries. However, I'm not lawyer and not qualified to give legal
advice.